+++ packages/bindings/examples/schema.rs
@@ -3,7 +3,10 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use provwasm_std::{Attributes, Marker, Names, ProvenanceMsg, ProvenanceQuery, ProvenanceRoute};
+use provwasm_std::{
+    Attributes, Marker, Names, ProvenanceMsg, ProvenanceQuery, ProvenanceRoute, Records, Scope,
+    Sessions,
+};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
@@ -14,6 +17,9 @@ fn main() {
     export_schema(&schema_for!(Attributes), &out_dir);
     export_schema(&schema_for!(Marker), &out_dir);
     export_schema(&schema_for!(Names), &out_dir);
+    export_schema(&schema_for!(Scope), &out_dir);
+    export_schema(&schema_for!(Sessions), &out_dir);
+    export_schema(&schema_for!(Records), &out_dir);
 
     export_schema(&schema_for!(ProvenanceMsg), &out_dir);
     export_schema(&schema_for!(ProvenanceQuery), &out_dir);