+++ packages/mocks/src/lib.rs
@@ -2,12 +2,14 @@
 mod attribute;
 mod common;
 mod marker;
+mod metadata;
 mod name;
 mod querier;
 
 pub use attribute::AttributeQuerier;
 pub use common::must_read_binary_file;
 pub use marker::MarkerQuerier;
+pub use metadata::MetadataQuerier;
 pub use name::NameQuerier;
 
 pub use querier::{mock_dependencies, mock_dependencies_with_balances, ProvenanceMockQuerier};