+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased changes
 
-* None
+* Add metadata module query support for scopes, sessions, and records.
+* Add missing attribute value types to fix de-serialization errors.
 
 ## Releases